* commit 'e959bdc036035496ab694fe6ba225fe0917f6497': (429 commits)
  Remove extraneous logger.info and change a warning to logger.warning
  Make the prereq test more flexible and correct
  Implement a much simpler script_regression_test tests
  Partial implementation of the prereq test. Still need to find out a generic way of verifying that the dependant job actually depends on the other one
  Revert change from afterok to afterany; this should only be done if the user specifically requests it
  Implements the prereq argument for case.submit, allowing the user to specify jobs which should finish (not necessarily successfully) before running the current job
  Update ChangeLog
  Change case.st_archive dependency string to not include strings indicating unsupported logic. Fix cobalt depend_string. Add depend_separator field to support LSF
  update mpt and pnetcdf on cheyenne
  Update description for TOTALPES
  Include spare nodes in COST_PES
  Update ChangeLog
  Add mapping files for T42_T42_mg17
  Significant refactor to aprun
  Fix aprun unit test.
  Fix COST_PES and TOTALPES for titan (aprun)
  Fix for multiple couplers. Also fix typo regression and simplify code somewhat
  update comment
  empty proxy tag in machines file invalid, removed from anlworkstation
  response to review
  ...

+++ scripts/lib/CIME/SystemTests/erp.py
@@ -34,10 +34,10 @@ class ERP(RestartTest):
             nthreads  = self._case1.get_value("NTHRDS_{}".format(comp))
             rootpe    = self._case1.get_value("ROOTPE_{}".format(comp))
             if ( nthreads > 1 ):
-                self._case.set_value("NTHRDS_{}".format(comp), nthreads/2)
+                self._case.set_value("NTHRDS_{}".format(comp), int(nthreads/2))
             if ( ntasks > 1 ):
-                self._case.set_value("NTASKS_{}".format(comp), ntasks/2)
-                self._case.set_value("ROOTPE_{}".format(comp), rootpe/2)
+                self._case.set_value("NTASKS_{}".format(comp), int(ntasks/2))
+                self._case.set_value("ROOTPE_{}".format(comp), int(rootpe/2))
 
         RestartTest._case_two_setup(self)
         # Note, some components, like CESM-CICE, have
